+++ src/doomstat.c
@@ -258,6 +258,7 @@ int comp_switchsource;
 int comp_cgundblsnd;
 int comp_cgunnersfx;
 int comp_flamst;
+int comp_keynoway;
 int comp_godface;
 int comp_deadoof;
 int comp_unusedpals;