+++ src/doomstat.h
@@ -651,6 +651,7 @@ extern int comp_switchsource;
 extern int comp_cgundblsnd;
 extern int comp_cgunnersfx;
 extern int comp_flamst;
+extern int comp_keynoway;
 extern int comp_godface;
 extern int comp_deadoof;
 extern int comp_unusedpals;